+++ lenskit/lenskit/util/test.py
@@ -26,7 +26,8 @@ from lenskit.algorithms.basic import PopScore
 from lenskit.algorithms.ranking import PlackettLuce
 from lenskit.batch import recommend
 from lenskit.crossfold import simple_test_pair
-from lenskit.data.dataset import Dataset, LazyDataset, from_interactions_df
+from lenskit.data import Dataset, from_interactions_df
+from lenskit.data.lazy import LazyDataset
 from lenskit.data.movielens import load_movielens, load_movielens_df
 from lenskit.math.sparse import torch_sparse_from_scipy
